requirements.txt
Django>=5.2.8,<6.0
djangorestframework>=3.16.1,<4.0

psycopg2-binary
drf-yasg==1.21.11
 
djangorestframework-simplejwt
